feat(core): Add function to determine if the active user is an admin
Maintainers can specify their admin groups in `src/admin-groups`
