+++ Cargo.toml
@@ -30,3 +30,4 @@ futures = "0.1.28"
 log = "0.4.8"
 shrinkwraprs = "0.2.1"
 stream-cancel = "0.4.4"
+users = "0.9.1"
