+++ gtk/Cargo.toml.in
@@ -25,5 +25,3 @@ fern = "0.5.8"
 log = "0.4.8"
 clap = "2.33.0"
 yansi = "0.5.0"
-stream-cancel = "0.4.4"
-users = "0.9.1"
