+++ src/admin-groups
@@ -1,0 +1,2 @@
+adm
+sudo
